refactor: TuripStreamResult 및 StreamFailureAction의 패키지 위치를 data에서 domain으로 이동
