+++ android/app/src/main/java/com/on/turip/domain/turip/result/StreamFailureAction.kt
@@ -1,4 +1,4 @@
-package com.on.turip.data.result
+package com.on.turip.domain.turip.result
 
 enum class StreamFailureAction {
     Retry,
